move newly adding tests to separate test to avoid multi restart repose
